linux-headers: Add ARM KVM headers (not for upstream)
This commit adds the ARM KVM headers. This is not to go to QEMU
upstream -- the correct path there is that the KVM code will be
committed to a mainline upstream kernel, and then upstream QEMU
can do a bulk header update from the upstream kernel, which will
allow us to drop this temporary commit.

This is the result of running update-headers on Christoffer's
kvm-arm-master branch commit 6024647. It includes the new
interrupt delivery ABI and the switch to ONE_REG for core
and coprocessor registers.

It currently includes some minor x86 header changes which
hopefully will have made it into QEMU upstream by the time
we submit this for merging.