hw/pci-host: add a generic PCI host
Add a generic PCI host controller for virtual platforms. This is for ARM
virtual platforms at the moment, but has nothing ARM specific. This
matches the generic PCI host driver added in the 3.16 kernel.

Functioning with OHCI (usb disk), but not LSI SCSI which doesn't find a
disk.

TODO:
- The memory region size and alias offset for the io and memory space
are hardcoded and need to be set from the machine and must match the
values in the DT ranges. Probably need to make these properties.
- More space for PCI memory space needed?
- MSI support - Needs GICv2M or GICv3 support.

Signed-off-by: Rob Herring <rob.herring@linaro.org>
Signed-off-by: Alvise Rigo <a.rigo@virtualopensystems.com>